requirements.txt
easyocr
langchain
langchain-core
langchain-community
langchain-groq
streamlit
python-dotenv